Separate functions from index.js file.
